refactor(wave7a-step3): close authorizer split behind thin facade
